# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API za praćenje kompetencija",
        "description": "Dokumentacija obuhvata rute za upravljanje kompetencijama, verifikacijama i autentifikaciju korisnika.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni server"
        }
    ],
    "paths": {
...
        "/api/moderator/pending-verifications": {
            "get": {
                "tags": [
                    "Moderator"
                ],
                "summary": "Lista zahteva na čekanju (samo za moderatore)",
                "operationId": "ada41032c7c1792325595e26f433e5a3",
                "responses": {
                    "200": {
                        "description": "Uspešno dobavljena lista"
                    },
                    "403": {
                        "description": "Zabranjen pristup - niste moderator"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/moderator/verifications/{id}": {
            "get": {
                "tags": [
                    "Moderator"
                ],
                "summary": "Prikaz jedne specifične verifikacije",
                "operationId": "bbe9c2087b740509008577e07dc0f899",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalji verifikacije"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/moderator/verify/{id}": {
            "post": {
                "tags": [
                    "Moderator"
                ],
                "summary": "Odobravanje kompetencije",
                "operationId": "c9f476607326e359ecde80fca96b6c45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kompetencija uspešno odobrena"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/moderator/reject/{id}": {
            "post": {
                "tags": [
                    "Moderator"
                ],
                "summary": "Odbijanje kompetencije",
                "operationId": "b61542cd92fe8d6344ff06bb71b1a8d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "note": {
                                        "type": "string",
                                        "example": "Nedovoljno dokaza."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kompetencija odbijena"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/moderator/history": {
            "get": {
                "tags": [
                    "Moderator"
                ],
                "summary": "Istorija obrađenih zahteva moderatora",
                "operationId": "d5b9453bf6a51f2fd9ed401c2ddc578b",
                "responses": {
                    "200": {
                        "description": "Lista odobrenih i odbijenih zahteva"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/verifications-overview": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Pregled svih verifikacija (Admin)",
                "operationId": "92f7cecf07244f11198bc63c72253286",
                "responses": {
                    "200": {
                        "description": "Kompletna lista svih statusa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}